Change files to chosen solution and remove obsolete files
